# handler/user_handler.go
package handler

import (
	"net/http"

	"github.com/alvingxv/kanban-board-kelompok5/dto"
	"github.com/alvingxv/kanban-board-kelompok5/entity"
	"github.com/alvingxv/kanban-board-kelompok5/pkg/errs"
	"github.com/alvingxv/kanban-board-kelompok5/service"
	"github.com/gin-gonic/gin"
)

type userHandler struct {
	userService service.UserService
}

func NewUserHandler(userService service.UserService) userHandler {
	return userHandler{
		userService: userService,
	}
}

// UserRegister godoc
// @Tags users
// @Description Create New User Data
// @ID create-new-user
// @Accept json
// @Produce json
// @Param RequestBody body dto.RegisterRequest true "request body json"
// @Success 201 {object} dto.RegisterResponse
// @Router /users/register [post]
func (uh *userHandler) Register(ctx *gin.Context) {
	var newUserRequest dto.RegisterRequest

	if err := ctx.ShouldBindJSON(&newUserRequest); err != nil {
		errBindJson := errs.NewUnprocessibleEntityError("invalid request body")

		ctx.JSON(errBindJson.Status(), errBindJson)
		return
	}

	result, err := uh.userService.Register(newUserRequest)

	if err != nil {
		ctx.JSON(err.Status(), err)
		return
	}

	ctx.JSON(http.StatusCreated, result)

}

func (uh *userHandler) Login(ctx *gin.Context) {
	var loginRequest dto.LoginRequest

	if err := ctx.ShouldBindJSON(&loginRequest); err != nil {
		errBindJson := errs.NewUnprocessibleEntityError("invalid request body")

		ctx.JSON(errBindJson.Status(), errBindJson)
		return
	}

	result, err := uh.userService.Login(loginRequest)

	if err != nil {
		ctx.JSON(err.Status(), err)
		return
	}
	ctx.JSON(http.StatusOK, result)
}

func (uh *userHandler) UpdateUser(ctx *gin.Context) {
	userData := ctx.MustGet("userData").(*entity.User)

	var updateRequest dto.UpdateRequest

	if err := ctx.ShouldBindJSON(&updateRequest); err != nil {
		errBindJson := errs.NewUnprocessibleEntityError("invalid request body")

		ctx.JSON(errBindJson.Status(), errBindJson)
		return
	}

	result, err := uh.userService.UpdateUser(updateRequest, userData)

	if err != nil {
		ctx.JSON(err.Status(), err)
		return
	}

	ctx.JSON(http.StatusOK, result)

}

func (uh *userHandler) DeleteUser(ctx *gin.Context) {
	userData := ctx.MustGet("userData").(*entity.User)

	err := uh.userService.DeleteUser(userData.ID)

	if err != nil {
		ctx.JSON(err.Status(), err)
		return
	}

	ctx.JSON(http.StatusOK, &gin.H{
		"Message": "Your Account has been successfully deleted",
	})

}